+++ pkg/stringsx/stringsx_test.go
@@ -12,24 +12,37 @@
 // OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
 // PERFORMANCE OF THIS SOFTWARE.
 
-// Package strutil holds small, dependency-free string helpers shared
-// across packages.
-package strutil
+package stringsx_test
 
-import "strings"
+import (
+	"testing"
 
-// NormalizeAlnum lowercases s and keeps only ASCII letters and digits, so
-// vendor names and domains can be compared free of spacing, punctuation,
-// and casing differences (e.g. "Letaido" and "letaido.com" both reduce to
-// a comparable form).
-func NormalizeAlnum(s string) string {
-	var b strings.Builder
+	"github.com/stretchr/testify/assert"
+	"go.probo.inc/probo/pkg/stringsx"
+)
 
-	for _, r := range strings.ToLower(s) {
-		if (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9') {
-			b.WriteRune(r)
-		}
+func TestNormalizeAlnum(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name     string
+		input    string
+		expected string
+	}{
+		{name: "letters lowercased", input: "Letaido", expected: "letaido"},
+		{name: "strips punctuation and spaces", input: "letaido.com - Inc", expected: "letaidocominc"},
+		{name: "keeps digits", input: "auth0", expected: "auth0"},
+		{name: "empty", input: "", expected: ""},
+		{name: "only punctuation", input: "-_.:", expected: ""},
 	}
 
-	return b.String()
+	for _, tt := range tests {
+		t.Run(
+			tt.name,
+			func(t *testing.T) {
+				t.Parallel()
+				assert.Equal(t, tt.expected, stringsx.NormalizeAlnum(tt.input))
+			},
+		)
+	}
 }
